Move "api-boot-plugin-datasource-switch" to "api-boot"
