--- api-boot-project/api-boot-plugins/api-boot-plugin-datasource-switch/pom.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>api-boot-plugins</artifactId>
-        <groupId>org.minbox.framework</groupId>
-        <version>${revision}</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>api-boot-plugin-datasource-switch</artifactId>
-    <description>
-        ApiBoot DataSource Switch
-        数据源切换：
-        1. 多数据源配置
-        2. 数据源路由器实现
-        3. 配置使用指定数据源
-        4. 动态使用指定数据源
-        5. 动态创建数据源
-    </description>
-    <dependencies>
-        <!--ApiBoot Plugin-->
-        <dependency>
-            <groupId>org.minbox.framework</groupId>
-            <artifactId>api-boot-plugin</artifactId>
-        </dependency>
-        <!--HikariCP DataSource-->
-        <dependency>
-            <groupId>com.zaxxer</groupId>
-            <artifactId>HikariCP</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <!--Druid DataSource-->
-        <dependency>
-            <groupId>com.alibaba</groupId>
-            <artifactId>druid</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-jdbc</artifactId>
-            <optional>true</optional>
-        </dependency>
-
-    </dependencies>
-</project>
