+++ api-boot-project/api-boot-plugins/pom.xml
@@ -22,7 +22,6 @@
         <module>api-boot-plugin-security</module>
         <module>api-boot-plugin-oauth</module>
         <module>api-boot-plugin-quartz</module>
-        <module>api-boot-plugin-datasource-switch</module>
         <module>api-boot-plugin-resource-load</module>
         <module>api-boot-plugin-http-converter</module>
         <module>api-boot-plugin-message-push</module>
