+++ api-boot-project/api-boot/src/main/java/org/minbox/framework/api/boot/datasource/support/ApiBootHikariDataSource.java
@@ -1,9 +1,9 @@
-package org.minbox.framework.api.boot.plugin.datasource.support;
+package org.minbox.framework.api.boot.datasource.support;
 
 import com.zaxxer.hikari.HikariDataSource;
 import org.minbox.framework.api.boot.common.exception.ApiBootException;
-import org.minbox.framework.api.boot.plugin.datasource.ApiBootDataSource;
-import org.minbox.framework.api.boot.plugin.datasource.config.DataSourceHikariConfig;
+import org.minbox.framework.api.boot.datasource.ApiBootDataSource;
+import org.minbox.framework.api.boot.datasource.config.DataSourceHikariConfig;
 
 import javax.sql.DataSource;
